+++ finetuner/run.py
@@ -1,5 +1,13 @@
 from finetuner.client import FinetunerV1Client
-from finetuner.constants import ARTIFACTS_DIR, FINISHED, STATUS
+from finetuner.constants import (
+    ARTIFACT_ID,
+    ARTIFACTS_DIR,
+    CREATED,
+    FAILED,
+    STARTED,
+    STATUS,
+)
+from finetuner.exception import RunFailedError, RunInProgressError
 from finetuner.hubble import download_artifact
 
 
@@ -29,6 +37,7 @@ class Run:
         self._config = config
         self._created_at = created_at
         self._description = description
+        self._run = self._get_run()
 
     @property
     def name(self) -> str:
@@ -37,6 +46,12 @@ class Run:
     @property
     def config(self) -> dict:
         return self._config
+
+    def _get_run(self) -> dict:
+        """Get Run object as dict."""
+        return self._client.get_run(
+            experiment_name=self._experiment_name, run_name=self._name
+        )
 
     def status(self) -> dict:
         """Run status.
@@ -56,18 +71,41 @@ class Run:
             experiment_name=self._experiment_name, run_name=self._name
         )
 
+    def _check_run_status(self):
+        status = self.status()[STATUS]
+        if status in [CREATED, STARTED]:
+            raise RunInProgressError(
+                'The run needs to be finished in order to save the artifact.'
+            )
+        if status == FAILED:
+            raise RunFailedError(
+                'The run failed, please check the `logs` for detailed information.'
+            )
+
     def save_artifact(self, directory: str = ARTIFACTS_DIR) -> str:
         """Save artifact if the run is finished.
 
         :param directory: Directory where the artifact will be stored.
         :returns: A string object that indicates the download path.
         """
-        if self.status()[STATUS] != FINISHED:
-            raise Exception('The run needs to be finished in order to save the model.')
-
+        self._check_run_status()
         return download_artifact(
             client=self._client,
-            experiment_name=self._experiment_name,
+            artifact_id=self._run[ARTIFACT_ID],
             run_name=self._name,
             directory=directory,
         )
+
+    @property
+    def artifact_id(self):
+        """Get artifact id from the run.
+
+        An artifact in finetuner contains fine-tuned model and its metadata.
+        Such as preprocessing function, collate function. This id could be useful
+        if you want to directly pull the artifact from the cloud storage, such as
+        using `FinetunerExecutor`.
+
+        :return: Artifact id as string object.
+        """
+        self._check_run_status()
+        return self._run[ARTIFACT_ID]